Compute filterRelation inversion outside events
